FIX: img delete on cancel

+++ src/utils/deleteFile.ts
@@ -2,13 +2,14 @@ import { ref, getStorage, deleteObject } from "firebase/storage";
 import { DocumentData, deleteDoc, doc } from "firebase/firestore";
 import { projectFirestore } from "../firebase/config";
 
-const deleteFile = async (file: DocumentData | null, collection: string) => {
-  if (!file) {
+export const deleteFileFromStorage = async (url: string) => {
+  if (!url) {
     console.log("Please choose a file first!");
     return;
   }
+
   const storage = getStorage();
-  const storageRef = ref(storage, file.url);
+  const storageRef = ref(storage, url);
 
   deleteObject(storageRef)
     .then(() => {
@@ -17,7 +18,15 @@ const deleteFile = async (file: DocumentData | null, collection: string) => {
     .catch((error) => {
       console.log(error);
     });
+};
 
+const deleteFile = async (file: DocumentData | null, collection: string) => {
+  if (!file) {
+    console.log("Please choose a file first!");
+    return;
+  }
+
+  deleteFileFromStorage(file.url);
   await deleteDoc(doc(projectFirestore, collection, file.id))
     .then(() => console.log("database deleted"))
     .catch((error) => console.log(error));
